03.29 구현

+++ daker/src/main/java/com/daker/domain/dto/request/ArticleRequestDTO.java
@@ -25,7 +25,7 @@ public class ArticleRequestDTO {
     @AllArgsConstructor
     @NoArgsConstructor
     public static class lookingForDTO {
-        long positionId;
+        int positionId;
         int headCount;
     }
 }
